fix: move out of hooks
