+++ jest.setup.tsx
@@ -94,8 +94,6 @@ jest.mock("next/navigation", () => {
   };
 });
 
-jest.mock("@/hooks/usePathServerSide", () => jest.fn());
-
 jest.mock("@/data/store", () => ({
   useStore: jest.fn(),
 }));
