+++ src/app/[locale]/(user-only)/layout.tsx
@@ -1,10 +1,10 @@
-import useProfileRedirect from "@/hooks/useProfileRedirect";
 import { PageContainer } from "@/modules";
 import ApplicationUser from "@/organisms/ApplicationUser";
+import { redirectProfile } from "@/utils/router";
 import { PropsWithChildren } from "react";
 
 export default async function Layout({ children }: PropsWithChildren) {
-  await useProfileRedirect();
+  await redirectProfile();
 
   return (
     <ApplicationUser>
