--- src/hooks/usePathServerSide/index.ts
@@ -1,3 +1,0 @@
-import { usePathServerSide } from "./usePathServerSide";
-
-export default usePathServerSide;
