--- src/hooks/useProfileRedirect/index.ts
@@ -1,3 +1,0 @@
-import useProfileRedirect from "./useProfileRedirect";
-
-export default useProfileRedirect;
