+++ src/utils/router.test.ts
@@ -1,6 +1,4 @@
 import { UserGroup } from "@/consts/user";
-import { ROLES, ROLES_STATE } from "../consts/roles";
-import { isRoleValid } from "./roles";
 import { needsLoggedInPermissions } from "./router";
 
 describe("needsLoggedInPermissions", () => {
