+++ src/utils/router.ts
@@ -1,4 +1,10 @@
+import { ROUTES } from "@/consts/router";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
+import getMe from "../services/auth/getMe";
 import { RouteConfig, Routes } from "../types/router";
+import { getAccessToken } from "./auth";
+import { getProfileRedirectPath, redirectToPath } from "./redirects";
 
 function needsLoggedInPermissions(routes: Routes, pathname: string | null) {
   if (!pathname) return false;
@@ -10,4 +16,29 @@ function needsLoggedInPermissions(routes: Routes, pathname: string | null) {
   });
 }
 
-export { needsLoggedInPermissions };
+function getPathServerSide(): string | null {
+  const head = headers();
+  return head.get("x-current-path");
+}
+
+async function redirectProfile() {
+  const pathname = getPathServerSide();
+
+  const accessToken = await getAccessToken();
+
+  if (!accessToken && needsLoggedInPermissions(ROUTES, pathname)) {
+    redirect(ROUTES.homepage.path);
+  }
+
+  if (accessToken && pathname) {
+    const response = await getMe({
+      suppressThrow: true,
+    });
+
+    if (response.status === 200) {
+      redirectToPath(getProfileRedirectPath(response.data), pathname);
+    }
+  }
+}
+
+export { getPathServerSide, needsLoggedInPermissions, redirectProfile };
